+++ lib/nfc_wallet_suppression.dart
@@ -1,5 +1,6 @@
 library;
 
 export 'src/nfc_wallet_suppression.dart';
+export 'src/nfc_wallet_suppression_platform_interface.dart';
 export 'src/nfc_wallet_suppression_status.dart';
 export 'src/nfc_wallet_suppression_stub.dart';